refactor(account): credential_manager 改名 account_manager，移除旧单账号函数
- credential_manager.py → account_manager.py
- 删除 load_credentials/save_credentials/get_credentials_filepath
- main_window 改用 load_accounts/save_accounts
- 移除迁移逻辑，旧用户升级后需重新登录一次
